+++ src/components/ItaliaTheme/Blocks/Accordion/Block/EditBlock.jsx
@@ -16,8 +16,7 @@ import {
 import { Button } from 'design-react-kit/dist/design-react-kit';
 import { Icon } from '@italia/components/ItaliaTheme';
 
-// import { TextEditorWidget } from '@italia/components/ItaliaTheme';
-import TextEditorWidget from '../TextEditorWidget';
+import { TextEditorWidget } from '@italia/components/ItaliaTheme';
 
 const messages = defineMessages({
   titlePlaceholder: {
@@ -111,6 +110,8 @@ class EditBlock extends SubblockEdit {
                   messages.titlePlaceholder,
                 )}
                 nextFocus="text"
+                prevFocus="text"
+                disableMoveToNearest={true}
                 setFocus={(f) => {
                   this.setState({ focusOn: f });
                 }}
@@ -138,7 +139,9 @@ class EditBlock extends SubblockEdit {
                   placeholder={this.props.intl.formatMessage(
                     messages.textPlaceholder,
                   )}
+                  disableMoveToNearest={true}
                   prevFocus="title"
+                  nextFocus="title"
                   setFocus={(f) => {
                     this.setState({ focusOn: f });
                   }}