+++ src/config/italiaConfig.js
@@ -50,6 +50,10 @@ import faTwitter from '@italia/icons/twitter-brands.svg';
 import TwitterPostsView from '@italia/components/ItaliaTheme/Blocks/TwitterPosts/View';
 import TwitterPostsEdit from '@italia/components/ItaliaTheme/Blocks/TwitterPosts/Edit';
 
+import iconBlocksSVG from '@plone/volto/icons/apps.svg';
+import IconBlocksView from '@italia/components/ItaliaTheme/Blocks/IconBlocks/View';
+import IconBlocksEdit from '@italia/components/ItaliaTheme/Blocks/IconBlocks/Edit';
+
 import CharCounterDescriptionWidget from '@italia/components/ItaliaTheme/manage/Widgets/CharCounterDescriptionWidget';
 import PageView from '@italia/components/ItaliaTheme/View/PageView/PageView';
 import NewsItemView from '@italia/components/ItaliaTheme/View/NewsItemView/NewsItemView';
@@ -72,6 +76,9 @@ import CardWithImageTemplateOptions from '@italia/components/ItaliaTheme/Blocks/
 
 import SmallBlockLinksTemplate from '@italia/components/ItaliaTheme/Blocks/Listing/SmallBlockLinksTemplate';
 import SmallBlockLinksTemplateSkeleton from '@italia/components/ItaliaTheme/Blocks/Listing/TemplatesSkeletons/SmallBlockLinksTemplateSkeleton';
+
+import SquaresImageTemplate from '@italia/components/ItaliaTheme/Blocks/Listing/SquaresImageTemplate';
+import SquaresImageTemplateSkeleton from '@italia/components/ItaliaTheme/Blocks/Listing/TemplatesSkeletons/SquaresImageTemplateSkeleton';
 
 import CompleteBlockLinksTemplate from '@italia/components/ItaliaTheme/Blocks/Listing/CompleteBlockLinksTemplate';
 import CompleteBlockLinksTemplateSkeleton from '@italia/components/ItaliaTheme/Blocks/Listing/TemplatesSkeletons/CompleteBlockLinksTemplateSkeleton';
@@ -421,6 +428,11 @@ export default function applyConfig(voltoConfig) {
           skeleton: RibbonCardTemplateSkeleton,
           templateOptions: RibbonCardTemplateOptions,
         },
+        squaresImageTemplate: {
+          label: 'Quadratoni con immagine',
+          template: SquaresImageTemplate,
+          skeleton: SquaresImageTemplateSkeleton,
+        },
         mapTemplate: {
           label: 'Mappa',
           template: MapTemplate,
@@ -475,6 +487,21 @@ export default function applyConfig(voltoConfig) {
       },
       listing_bg_colors: [], //{name:'blue', label:'Blu'},{name:'light-blue', label:'Light blue'},{name:'sidebar-background', label:'Grey'}
       listing_items_colors: [], //{name:'blue', label:'Blu'},{name:'light-blue', label:'Light blue'},{name:'sidebar-background', label:'Grey'}
+    },
+    iconBlocks: {
+      id: 'iconBlocks',
+      title: 'Blocchi con icone',
+      icon: iconBlocksSVG,
+      group: 'text',
+      view: IconBlocksView,
+      edit: IconBlocksEdit,
+      restricted: false,
+      mostUsed: true,
+      security: {
+        addPermission: [],
+        view: [],
+      },
+      sidebarTab: 1,
     },
     rssBlock,
   };